requirements.txt
pymupdf 
sentence-transformers 
faiss-cpu 
langchain_groq
flask
langsmith
python-dotenv
langchain_core
flask-cors
gunicorn